read_file instead of raw_data and raw_state
Reason: in the next commits we'll have also a "log" file that should be
just treated the same way, and in the not-far future we'll also probably have
"stdout"/"stderr" files that also won't be any different.

Also: probably CLI command also should be unified, but I'm not
super-sure, so this can wait.